// ProyectoIIB_Flores_Romero/Frontend/Contructora/src/app/propiedad.interface.ts
export interface Propiedad{
  ID_PROPIEDAD?: number;
  NOMBRE_PROP?: string;
  DIMENSION_TOTAL_PRO?: string;
  ID_PROYECTO?: number;
  PLANO_PROP?: string;
  PRECIO_PROP?: number;
  NUMERO_HABITACIONES_PROP?: number;
  NUMERO_BANOS_PROP?: number;
  NUMERO_ESTACIONAMIENTOS_PROP?: number;
  TIPO_PROP?: string;
  UBICACION_PROY?: string;

}